backend/requirements-production.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5

# API Documentation
drf-spectacular==0.26.5

# Database
psycopg2-binary==2.9.9

# Image Processing
Pillow==10.1.0

# PDF Generation
reportlab==4.0.7

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
hypothesis==6.92.1

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Environment Variables
python-decouple==3.8

# CORS
django-cors-headers==4.3.1

# Logging
python-json-logger==2.0.7

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# Async Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# NLP (for bot)
spacy==3.7.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Production dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
sentry-sdk==1.39.1
